added the compilation suit for keccak
